Add files via upload
Newest Farm Project

The goal is to design an Excel workbook to accomplish a few goals.

1. Put the truckloads into a worksheet in a consistent and clean way. This will make analysis easier in the future. 

2. Pull the information from the Load sheet to the Reports sheet. This will give the farmer a total yield per field/per crop/per year. It will also show him the sum of crop sales for the same. 

This is a simple project with a user friendly interface that the farmer can use for years to come. It has all the features he wants, and none he doesn't. He prefers the "KISS" method when it comes to computers.